backend/requirements.txt
Flask
flask-cors
flask-pymongo
pymongo[srv]
python-dotenv
cloudinary
bson
